+++ packages/jerni/src/dev-cli/syncReadableEventsToBinaryFile.ts
